feat(platform): full Windows MSVC support — compiler, stdlib, AOT, VCS
Complete Windows x86_64-pc-windows-msvc support across all layers:

LLVM Build Chain (7 files):
- All build.rs: llvm-config.exe detection, .lib name parsing, C++20
- verum_tblgen: NDEBUG for dump() compat, system-lib .lib suffix strip
- target.c: LLVMConstMul/NSWMul/NUWMul shims (removed in LLVM 17+)
- verum_mlir: i32 enum repr for LLVM 21 MLIR C API changes

V-LLSI No-Libc AOT (3 files):
- platform_ir.rs: VirtualAlloc/VirtualFree/WriteFile/ExitProcess
- mainCRTStartup entry point for Windows PE without CRT
- link.rs: /SUBSYSTEM:CONSOLE, /OPT:REF, /OPT:ICF
- lld/mod.rs: COFF /LIBPATH, /OUT, .lib, /DEBUG:NONE

Core Stdlib (10 .vr files):
- net/tcp.vr, udp.vr: separate Linux/Windows mount (winsock2)
- async/channel.vr: Windows futex mount
- base/env.vr: PATH separator ';' on Windows
- term/event/source.vr: blocking_read, has_data via Win32 API
- term/raw/screen.vr: write_to_fd, read_from_fd via WriteFile/ReadFile
- sys/windows/winsock2.vr: POSIX-compatible socket wrappers

VCS Test Infrastructure (9 files):
- Single verum binary (verum run --interp/--aot)
- Windows Job Object memory limits, GetProcessMemoryInfo RSS
- NTSTATUS crash detection in fuzzer
- Glob path normalization, Windows env vars

Compiler/Runtime (6 files):
- LTO stub on MSVC (single-pass linker limitation)
- libc::c_void → std::ffi::c_void, extern_abi_to_convention
- 16 MB stack for deep recursive compiler structures
- windows-sys deps for vtest/vbench

Verified: 1025+ unit tests pass, 0 failures on Windows MSVC.

+++ crates/llvm/verum_mlir/src/diagnostic/severity.rs
@@ -13,17 +13,17 @@ pub enum DiagnosticSeverity {
     Warning,
 }
 
-impl TryFrom<u32> for DiagnosticSeverity {
+impl TryFrom<i32> for DiagnosticSeverity {
     type Error = Error;
 
-    fn try_from(severity: u32) -> Result<Self, Error> {
+    fn try_from(severity: i32) -> Result<Self, Error> {
         #[allow(non_upper_case_globals)]
         Ok(match severity {
             MlirDiagnosticSeverity_MlirDiagnosticError => Self::Error,
             MlirDiagnosticSeverity_MlirDiagnosticNote => Self::Note,
             MlirDiagnosticSeverity_MlirDiagnosticRemark => Self::Remark,
             MlirDiagnosticSeverity_MlirDiagnosticWarning => Self::Warning,
-            _ => return Err(Error::UnknownDiagnosticSeverity(severity)),
+            _ => return Err(Error::UnknownDiagnosticSeverity(severity as u32)),
         })
     }
 }
